Major refactoring. Added tick support when graphing.
