+++ .gitignore
@@ -58,6 +58,18 @@ Makefile
 cmake_install.cmake 
 CMakeCache.txt
 CMakeFiles/
+src/CMakeFiles
+src/cmake_install.cmake
+src/CTestTestfile.cmake
+lvgl/CTestTestfile.cmake
+lv_drivers/CTestTestfile.cmake
+generated-includes/
+_deps/
+lib/
+Testing/
 .settings/
 .github/
+CTestTestfile.cmake
+DartConfiguration.tcl
+build/
 demo
