+++ .vscode/launch.json
@@ -5,10 +5,35 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Debug",
+            "name": "Debug Tests",
             "type": "cppdbg",
             "request": "launch",
-            "program": "${workspaceRoot}/demo",
+            "program": "${workspaceFolder}/tests/src/tests",
+            "args": [],
+            "stopAtEntry": false,
+            "cwd": "${fileDirname}",
+            "environment": [],
+            "externalConsole": false,
+            "MIMode": "gdb",
+            "setupCommands": [
+                {
+                    "description": "Enable pretty-printing for gdb",
+                    "text": "-enable-pretty-printing",
+                    "ignoreFailures": true
+                },
+                {
+                    "description":  "Set Disassembly Flavor to Intel",
+                    "text": "-gdb-set disassembly-flavor intel",
+                    "ignoreFailures": true
+                }
+            ],
+            "preLaunchTask": "build tests"
+        },
+        {
+            "name": "Debug SSGC",
+            "type": "cppdbg",
+            "request": "launch",
+            "program": "${workspaceRoot}/build/out/SSGC",
             "MIMode": "gdb",
             "setupCommands": [
                 {
@@ -22,7 +47,7 @@
                     "ignoreFailures": true
                 }
             ],
-            "target": "./demo",
+            "target": "${workspaceRoot}/build/out/SmartSymbolicGraphingCalculator",
             "cwd": "${workspaceRoot}",
             "preLaunchTask": "build",
             "miDebuggerPath": "/usr/bin/gdb"
