+++ CMakeLists.txt
@@ -1,15 +1,73 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 3.22.3)
-PROJECT (SmartSymbolicGraphingCalculator)
+PROJECT (SSGC)
 SET(CMAKE_CXX_STANDARD 14)
+set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 SET(CMAKE_C_FLAGS_DEBUG "-g -DDEBUG")
 SET(CMAKE_CXX_FLAGS_DEBUG "-g -DDEBUG")
 
+include(CTest)
+
+# include_directories(${CMAKE_SOURCE_DIR}/lv_drivers ${CMAKE_SOURCE_DIR}/lvgl ${CMAKE_SOURCE_DIR}/src)
+# include(${CMAKE_SOURCE_DIR}/tests/CMakeLists.txt)
+option(GIAC_ENABLE "Enable Giac library" ON)
+option(PI_ENABLE "Enable compiling/linking for the Raspberry PI Zero 2 W" OFF)
+option(LINUX_ENABLE "Enable compiling/linking for linux systems" ON)
+option(WINDOWS_ENABLE "Enable compiling/linking for windows" OFF)
+if(LINUX_ENABLE OR PI_ENABLE)
+    message(STATUS "Enabling Linux linkage...")
+    add_link_options(-lm -lpthread -lSDL2)
+endif()
+if(PI_ENABLE)
+    message(STATUS "Enabling Raspberry PI linkage...")
+    add_link_options(-li2c -lwiringPi)
+endif()
+if(ENABLE_GIAC)
+    message(STATUS "Enabling Giac linkage...")
+    add_link_options(-lgiac -lgmp)
+endif()
+if(WINDOW_ENABLE)
+    message(STATUS "Enabling Windows linkage...")
+    add_link_options(-lgdi32)
+endif()
+
+if(LINUX_ENABLE AND NOT PI_ENABLE)
+    set(ENABLE_LINUX 1)
+else()
+    set(ENABLE_LINUX 0)
+    set(ENABLE_PI 1)
+endif()
+if(PI_ENABLE)
+    set(ENABLE_PI 1)
+else()
+    set(ENABLE_PI 0)
+endif()
+if(GIAC_ENABLE)
+    set(ENABLE_GIAC 1)
+else()
+    set(ENABLE_GIAC 0)
+endif()
+if(WINDOWS_ENABLE)
+    set(ENABLE_WINDOWS 1)
+else()
+    set(ENABLE_WINDOWS 0)
+endif()
+configure_file(calc_conf.h.in calc_conf.h @ONLY)
+
+#message(${CMAKE_SYSTEM_NAME})
+
+#file(GLOB_RECURSE LVGL_SOURCE_FILES "lvgl/*.c")
+#file(GLOB_RECURSE LVGL_DRIVERS_SOURCE_FILES "lv_drivers/*.c")
+#add_link_options(-lm -lpthread -lSDL2 -lgmp -lgiac)
 include_directories(${CMAKE_SOURCE_DIR})
-set(SOURCE_FILES main.cpp Calculator.cpp graph.cpp graphing.cpp Keypad.cpp Settings.cpp tabs.cpp)
-add_link_options(-lm -lpthread -lSDL2 -lgmp -lgiac)
-file(GLOB_RECURSE LVGL_SOURCE_FILES "lvgl/*.c")
-file(GLOB_RECURSE LVGL_DRIVERS_SOURCE_FILES "lv_drivers/*.c")
+add_subdirectory(lv_drivers)
+add_subdirectory(lvgl)
+
+add_subdirectory(src)
+add_subdirectory(tests)
 
 
-ADD_EXECUTABLE(demo ${LVGL_DRIVERS_SOURCE_FILES} ${LVGL_SOURCE_FILES} ${SOURCE_FILES})
+
+
+
+#ADD_EXECUTABLE(demo ${LVGL_DRIVERS_SOURCE_FILES} ${LVGL_SOURCE_FILES} ${SOURCE_FILES})
