--- MCP23017.cpp
@@ -1,318 +1,0 @@
-#include "MCP23017.h"
-#if ENABLE_PI && !ENABLE_LINUX
-
- MCP23017::MCP23017(int bus, int address) {
-    kI2CBus = bus;           // I2C bus of Jetson (1 and 8 available on Xavier)
-    kI2CAddress = address ; // Address of MCP23017; defaults to 0x20
-    error = 0 ;
-}
-
-MCP23017::~MCP23017() {
-    closeI2C() ;
-}
-
-//helper function to replace bitRead function from Arduino library
-bool MCP23017::bitRead(uint8_t num, uint8_t index)
-{
-    return (num >> index) & 1;
-}
-
-
-//helper function to replace bitWrite function from Arduino library
-void MCP23017::bitWrite(uint8_t &var, uint8_t index, uint8_t bit)
-{
-    uint new_bit = 1 << index;
-    if(bit)
-    {
-        var = var | new_bit;
-    }
-    else {
-        new_bit = ~new_bit;
-        var = var & new_bit;
-    }
-}
-
-
-//open I2C communication
-bool MCP23017::openI2C()
-{
-    char fileNameBuffer[32];
-    sprintf(fileNameBuffer,"/dev/i2c-%d", kI2CBus);
-    kI2CFileDescriptor = open(fileNameBuffer, O_RDWR);
-    if (kI2CFileDescriptor < 0) {
-        // Could not open the file
-       error = errno ;
-       return false ;
-    }
-    if (ioctl(kI2CFileDescriptor, I2C_SLAVE, kI2CAddress) < 0) {
-        // Could not open the device on the bus
-        error = errno ;
-        return false ;
-    }
-    // set defaults!
-	// all inputs on port A and B
-	writeRegister(MCP23017_IODIRA,0b11111110);
-	writeRegister(MCP23017_IODIRB,0b11111110);
-    return true ;
-}
-
-//close I2C communication
-void MCP23017::closeI2C()
-{
-    if (kI2CFileDescriptor > 0) {
-        close(kI2CFileDescriptor);
-        // WARNING - This is not quite right, need to check for error first
-        kI2CFileDescriptor = -1 ;
-    }
-}
-
-
-
-/**
- * Bit number associated to a given Pin
- */
-uint8_t MCP23017::bitForPin(uint8_t pin){
-	return pin%8;
-}
-
-/**
- * Register address, port dependent, for a given PIN
- */
-uint8_t MCP23017::regForPin(uint8_t pin, uint8_t portAaddr, uint8_t portBaddr){
-	return(pin<8) ?portAaddr:portBaddr;
-}
-
-/**
- * Reads a given register
- */
-uint8_t MCP23017::readRegister(uint8_t addr)
-{
-    int toReturn = i2c_smbus_read_byte_data(kI2CFileDescriptor, addr);
-    if (toReturn < 0) {
-        printf("MCP23017 Read Byte error: %d",errno) ;
-        error = errno ;
-        toReturn = -1 ;
-    }
-    // For debugging
-    // printf("Device 0x%02X returned 0x%02X from register 0x%02X\n", kI2CAddress, toReturn, readRegister);
-    return toReturn ;
-}
-
-/**
- * Writes a given register
- */
-uint8_t MCP23017::writeRegister(uint8_t addr, uint8_t writeValue)
-{   // For debugging:
-    // printf("Wrote: 0x%02X to register 0x%02X \n",writeValue, writeRegister) ;
-
-    int toReturn = i2c_smbus_write_byte_data(kI2CFileDescriptor, addr, writeValue);
-    if (toReturn < 0) {
-        perror("Write to I2C Device failed");
-        error = errno ;
-        toReturn = -1 ;
-    }
-    return toReturn ;
-
-}
-
-/**
- * Reads a byte
- */
-uint8_t MCP23017::readByte()
-{
-    int toReturn = i2c_smbus_read_byte(kI2CFileDescriptor);
-    if (toReturn < 0) {
-        printf("MCP23017 Read Byte error: %d",errno) ;
-        error = errno ;
-        toReturn = -1 ;
-    }
-    // For debugging
-    // printf("Device 0x%02X returned 0x%02X from register 0x%02X\n", kI2CAddress, toReturn, readRegister);
-    return toReturn ;
-}
-
-/**
- * Writes a byte
- */
-uint8_t MCP23017::writeByte(uint8_t writeValue)
-{   // For debugging:
-    // printf("Wrote: 0x%02X to register 0x%02X \n",writeValue, writeRegister) ;
-    int toReturn = i2c_smbus_write_byte(kI2CFileDescriptor, writeValue);
-    if (toReturn < 0) {
-        printf("MCP23017 Write Byte error: %d",errno) ;
-        error = errno ;
-        toReturn = -1 ;
-    }
-    return toReturn ;
-}
-
-/**
- * Helper to update a single bit of an A/B register.
- * - Reads the current register value
- * - Writes the new register value
- */
-void MCP23017::updateRegisterBit(uint8_t pin, uint8_t pValue, uint8_t portAaddr, uint8_t portBaddr) {
-	uint8_t regValue;
-	uint8_t regAddr=regForPin(pin,portAaddr,portBaddr);
-	uint8_t bit=bitForPin(pin);
-	regValue = readRegister(regAddr);
-
-	// set the value for the particular bit
-	bitWrite(regValue,bit,pValue);
-
-	writeRegister(regAddr,regValue);
-}
-
-/**
- * Sets the pin mode to either INPUT or OUTPUT
- */
-void MCP23017::pinMode(uint8_t p, uint8_t d) {
-	updateRegisterBit(p,(d==INPUT),MCP23017_IODIRA,MCP23017_IODIRB);
-}
-
-/**
- * Reads all 16 pins (port A and B) into a single 16 bits variable.
- */
-uint16_t MCP23017::readGPIOAB() {
-	uint16_t ba = 0;
-	uint8_t a;
-
-	// read the current GPIO output latches
-	writeByte(MCP23017_GPIOA);
-
-	a = readByte();
-	ba = readByte();
-	ba <<= 8;
-	ba |= a;
-
-	return ba;
-}
-
-/**
- * Read a single port, A or B, and return its current 8 bit value.
- * Parameter b should be 0 for GPIOA, and 1 for GPIOB.
- */
-uint8_t MCP23017::readGPIO(uint8_t b) {
-
-	// read the current GPIO output latches
-	if (b == 0)
-		writeByte(MCP23017_GPIOA);
-	else {
-		writeByte(MCP23017_GPIOB);
-	}
-
-
-	uint8_t value = readByte();
-	return value;
-}
-
-/**
- * Writes all the pins in one go. This method is very useful if you are implementing a multiplexed matrix and want to get a decent refresh rate.
- */
-void MCP23017::writeGPIOAB(uint16_t ba) {
-
-	writeByte(MCP23017_GPIOA);
-	writeByte(ba & 0xFF);
-	writeByte(ba >> 8);
-
-}
-
-void MCP23017::digitalWrite(uint8_t pin, uint8_t d) {
-	uint8_t gpio;
-	uint8_t bit=bitForPin(pin);
-
-
-	// read the current GPIO output latches
-	uint8_t regAddr=regForPin(pin,MCP23017_OLATA,MCP23017_OLATB);
-	gpio = readRegister(regAddr);
-
-	// set the pin and direction
-	bitWrite(gpio,bit,d);
-
-	// write the new GPIO
-	regAddr=regForPin(pin,MCP23017_GPIOA,MCP23017_GPIOB);
-	writeRegister(regAddr,gpio);
-}
-
-void MCP23017::pullUp(uint8_t p, uint8_t d) {
-	updateRegisterBit(p,d,MCP23017_GPPUA,MCP23017_GPPUB);
-}
-
-bool MCP23017::digitalRead(uint8_t pin) {
-	uint8_t bit=bitForPin(pin);
-	uint8_t regAddr=regForPin(pin,MCP23017_GPIOA,MCP23017_GPIOB);
-	return (readRegister(regAddr) >> bit) & 0x1;
-}
-
-/**
- * Configures the interrupt system. both port A and B are assigned the same configuration.
- * Mirroring will OR both INTA and INTB pins.
- * Opendrain will set the INT pin to value or open drain.
- * polarity will set LOW or HIGH on interrupt.
- * Default values after Power On Reset are: (false, false, LOW)
- * If you are connecting the INTA/B pin to arduino 2/3, you should configure the interupt handling as FALLING with
- * the default configuration.
- */
-void MCP23017::setupInterrupts(uint8_t mirroring, uint8_t openDrain, uint8_t polarity){
-	// configure the port A
-	uint8_t ioconfValue=readRegister(MCP23017_IOCONA);
-	bitWrite(ioconfValue,6,mirroring);
-	bitWrite(ioconfValue,2,openDrain);
-	bitWrite(ioconfValue,1,polarity);
-	writeRegister(MCP23017_IOCONA,ioconfValue);
-
-	// Configure the port B
-	ioconfValue=readRegister(MCP23017_IOCONB);
-	bitWrite(ioconfValue,6,mirroring);
-	bitWrite(ioconfValue,2,openDrain);
-	bitWrite(ioconfValue,1,polarity);
-	writeRegister(MCP23017_IOCONB,ioconfValue);
-}
-
-/**
- * Set's up a pin for interrupt. uses arduino MODEs: CHANGE, FALLING, RISING.
- *
- * Note that the interrupt condition finishes when you read the information about the port / value
- * that caused the interrupt or you read the port itself. Check the datasheet can be confusing.
- *
- */
-void MCP23017::setupInterruptPin(uint8_t pin, uint8_t mode) {
-
-	// set the pin interrupt control (0 means change, 1 means compare against given value);
-	updateRegisterBit(pin,(mode!=CHANGE),MCP23017_INTCONA,MCP23017_INTCONB);
-	// if the mode is not CHANGE, we need to set up a default value, different value triggers interrupt
-
-	// In a RISING interrupt the default value is 0, interrupt is triggered when the pin goes to 1.
-	// In a FALLING interrupt the default value is 1, interrupt is triggered when pin goes to 0.
-	updateRegisterBit(pin,(mode==FALLING),MCP23017_DEFVALA,MCP23017_DEFVALB);
-
-	// enable the pin for interrupt
-	updateRegisterBit(pin,HIGH,MCP23017_GPINTENA,MCP23017_GPINTENB);
-
-}
-
-uint8_t MCP23017::getLastInterruptPin(){
-	uint8_t intf;
-
-	// try port A
-	intf=readRegister(MCP23017_INTFA);
-	for(int i=0;i<8;i++) if (bitRead(intf,i)) return i;
-
-	// try port B
-	intf=readRegister(MCP23017_INTFB);
-	for(int i=0;i<8;i++) if (bitRead(intf,i)) return i+8;
-
-	return MCP23017_INT_ERR;
-
-}
-uint8_t MCP23017::getLastInterruptPinValue(){
-	uint8_t intPin=getLastInterruptPin();
-	if(intPin!=MCP23017_INT_ERR){
-		uint8_t intcapreg=regForPin(intPin,MCP23017_INTCAPA,MCP23017_INTCAPB);
-		uint8_t bit=bitForPin(intPin);
-		return (readRegister(intcapreg)>>bit) & (0x01);
-	}
-
-	return MCP23017_INT_ERR;
-}
-#endif
