--- graphing.hxx
@@ -1,21 +1,0 @@
-#ifndef LV_DEMO_GRAPHING
-#define LV_DEMO_GRAPHING
-
-// #ifdef __cplusplus
-// extern "C" {
-// #endif
-
-#include "lv_drv_conf.h"
-#include "lvgl/lvgl.h"
-
-// #ifdef __cplusplus
-// }
-// #endif
-
-void create_graph(lv_obj_t *parent);
-
-lv_obj_t* get_functionTextArea();
-//void create_graph(void);
-
-
-#endif
