+++ src/include/tabs.hxx
@@ -1,0 +1,4 @@
+#include <stdio.h>
+#include <graphing.hxx>
+#include <Settings.hxx>
+lv_obj_t* lv_main_screen_tabs(void);
