--- tabs.hxx
@@ -1,4 +1,0 @@
-#include <stdio.h>
-#include "graphing.hxx"
-#include "Settings.hxx"
-lv_obj_t* lv_main_screen_tabs(void);
