+++ tests/src/CMakeLists.txt
@@ -1,0 +1,9 @@
+if(BUILD_TESTING)
+    add_executable(tests ticktest.cpp)
+    target_include_directories(tests PUBLIC ${CMAKE_SOURCE_DIR} ${CMAKE_SOURCE_DIR}/src/include)
+    set_target_properties(tests PROPERTIES CXX_EXTENSIONS OFF OUTPUT_NAME tests)
+    target_compile_features(tests PUBLIC cxx_std_14)
+    #target_compile_options(tests PRIVATE)
+    target_link_libraries(tests PRIVATE Catch2::Catch2WithMain)
+    add_test(NAME tests COMMAND tests)
+endif()
